Move BatterySummaryLarge.kt to batterysync/mobile
