+++ batterysync/mobile/src/androidMain/res/values/plurals.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+
+    <plurals name="battery_sync_last_updated_minutes">
+        <item quantity="one">Last updated %s minute ago</item>
+        <item quantity="other">Last updated %s minutes ago</item>
+    </plurals>
+
+</resources>
