+++ mobile/src/main/kotlin/com/boswelja/smartwatchextensions/batterysync/ui/BatterySyncSettingsHeader.kt
@@ -17,7 +17,6 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
 import com.boswelja.smartwatchextensions.R
-import com.boswelja.smartwatchextensions.common.BatteryIcon
 
 private const val HEADER_ASPECT_RATIO = 3f
