+++ mobile/src/main/res/values/plurals.xml
@@ -6,11 +6,6 @@
         <item quantity="other">%s permissions requested</item>
     </plurals>
 
-    <plurals name="battery_sync_last_updated_minutes">
-        <item quantity="one">Last updated %s minute ago</item>
-        <item quantity="other">Last updated %s minutes ago</item>
-    </plurals>
-
     <plurals name="app_manager_app_count">
         <item quantity="one">%s\napp</item>
         <item quantity="other">%s\napps</item>
